+++ ribs/visualize/__init__.py
@@ -14,6 +14,7 @@ object-oriented API, pass the `ax` parameter to these functions.
 .. autosummary::
     :toctree:
 
+    archive_histogram
     cvt_archive_3d_plot
     cvt_archive_heatmap
     grid_archive_heatmap
@@ -24,6 +25,7 @@ object-oriented API, pass the `ax` parameter to these functions.
     qdax_repertoire_heatmap
 """
 
+from ribs.visualize._archive_histogram import archive_histogram
 from ribs.visualize._cvt_archive_3d_plot import cvt_archive_3d_plot
 from ribs.visualize._cvt_archive_heatmap import cvt_archive_heatmap
 from ribs.visualize._grid_archive_heatmap import grid_archive_heatmap
@@ -38,6 +40,7 @@ from ribs.visualize._visualize_qdax import (
 )
 
 __all__ = [
+    "archive_histogram",
     "cvt_archive_3d_plot",
     "cvt_archive_heatmap",
     "grid_archive_heatmap",